refactor instructions into a parent module with a gate submodule
