+++ quil-py/src/instruction/calibration.rs
@@ -1,14 +1,9 @@
-// come back when:
-// [X] Instruction is defined
-// GateModifier is defined
-// Expression is defined
-// [X] Qubit is defined
-//
-
 use pyo3::{types::PyString, Py};
 use quil_rs::{
     expression::Expression,
-    instruction::{Calibration, GateModifier, Instruction, MeasureCalibrationDefinition, Qubit},
+    instruction::{
+        gate::GateModifier, Calibration, Instruction, MeasureCalibrationDefinition, Qubit,
+    },
 };
 use rigetti_pyo3::{impl_repr, impl_str, py_wrap_data_struct};
