+++ quil-rs/src/parser/command.rs
@@ -20,17 +20,20 @@ use nom::{
 };
 
 use crate::{
-    instruction::{Convert, GateSpecification, GateType, Include, PragmaArgument},
+    instruction::{
+        gate::{GateSpecification, GateType},
+        Convert, Include, PragmaArgument,
+    },
     parser::common::parse_variable_qubit,
 };
 
 use crate::instruction::{
-    Arithmetic, ArithmeticOperand, ArithmeticOperator, BinaryLogic, BinaryOperator, Calibration,
-    Capture, CircuitDefinition, Comparison, ComparisonOperator, Declaration, Delay, Exchange,
-    Fence, FrameDefinition, GateDefinition, Instruction, Jump, JumpUnless, JumpWhen, Label, Load,
-    MeasureCalibrationDefinition, Measurement, Move, Pragma, Pulse, Qubit, RawCapture, Reset,
-    SetFrequency, SetPhase, SetScale, ShiftFrequency, ShiftPhase, Store, UnaryLogic, UnaryOperator,
-    Waveform, WaveformDefinition,
+    gate::GateDefinition, Arithmetic, ArithmeticOperand, ArithmeticOperator, BinaryLogic,
+    BinaryOperator, Calibration, Capture, CircuitDefinition, Comparison, ComparisonOperator,
+    Declaration, Delay, Exchange, Fence, FrameDefinition, Instruction, Jump, JumpUnless, JumpWhen,
+    Label, Load, MeasureCalibrationDefinition, Measurement, Move, Pragma, Pulse, Qubit, RawCapture,
+    Reset, SetFrequency, SetPhase, SetScale, ShiftFrequency, ShiftPhase, Store, UnaryLogic,
+    UnaryOperator, Waveform, WaveformDefinition,
 };
 use crate::parser::instruction::parse_block;
 use crate::parser::InternalParserResult;
@@ -563,12 +566,15 @@ mod tests {
         Expression, ExpressionFunction, FunctionCallExpression, InfixExpression, InfixOperator,
         PrefixExpression, PrefixOperator,
     };
-    use crate::instruction::{GateDefinition, GateSpecification, PragmaArgument};
+    use crate::instruction::{
+        gate::{GateDefinition, GateSpecification},
+        PragmaArgument,
+    };
     use crate::parser::lexer::lex;
     use crate::{imag, real};
     use crate::{
         instruction::{
-            CircuitDefinition, Declaration, Gate, Instruction, Measurement, MemoryReference,
+            gate::Gate, CircuitDefinition, Declaration, Instruction, Measurement, MemoryReference,
             Pragma, Qubit, ScalarType, Vector,
         },
         make_test,
