+++ quil-rs/src/parser/gate.rs
@@ -22,7 +22,7 @@ use super::{
     expression::parse_expression,
     ParserErrorKind, ParserInput,
 };
-use crate::instruction::Gate;
+use crate::instruction::gate::Gate;
 use crate::parser::InternalParserResult;
 
 /// Parse a gate instruction.
@@ -52,7 +52,10 @@ pub(crate) fn parse_gate<'a>(input: ParserInput<'a>) -> InternalParserResult<'a,
 mod test {
     use super::parse_gate;
     use crate::expression::Expression;
-    use crate::instruction::{Gate, GateModifier, Instruction, Qubit};
+    use crate::instruction::{
+        gate::{Gate, GateModifier},
+        Instruction, Qubit,
+    };
     use crate::make_test;
     use crate::parser::lexer::lex;
